requirements.txt
django
django-tables2
django-crispy-forms